feature\fix: Conf based output of object classes
Removes the jigsass-define-object mixin.

Previously, the `jigsass-define-object` mixin created a placeholder
selector for each object and object modifier, in each length and misc
breakpoints, and in every possible combination of them.

This allowed for their later extension using the `jigsass-object`
mixin while producing minimal css output with a predictable cascade.

Unfortunately, this approach proved problematic performance wise, and at
times caused the compiler to crash.

This commit addresses the issue by switching to generation of object
classes and their modifiers only if they are enabled in the the object's
configuration map.

Update tests
Update documentation.